+++ components/collector/src/source_collectors/performancetest_runner/performancetest_scalability.py
@@ -1,7 +1,6 @@
 """Performancetest-runner performancetest scalability collector."""
 
-from collector_utilities.exceptions import CollectorException
-from collector_utilities.type import Value
+from collector_utilities.type import Response, Value
 from model import SourceResponses
 
 from .base import PerformanceTestRunnerBaseClass
@@ -12,12 +11,8 @@ class PerformanceTestRunnerScalability(PerformanceTestRunnerBaseClass):
 
     async def _parse_value(self, responses: SourceResponses) -> Value:
         """Override to parse the scalability breaking point from the responses."""
-        trend_breaks = []
-        for response in responses:
-            breaking_point = int((await self._soup(response)).find(id="trendbreak_scalability").string)
-            if breaking_point == 100:
-                raise CollectorException(
-                    "No performance scalability breaking point occurred (breaking point is at 100%, expected < 100%)"
-                )
-            trend_breaks.append(breaking_point)
-        return str(min(trend_breaks))
+        return str(min([await self.__breaking_point(response) for response in responses]))
+
+    async def __breaking_point(self, response: Response) -> int:
+        """Parse the breaking point from the response."""
+        return int((await self._soup(response)).find(id="trendbreak_scalability").string)